requirements.txt
python-telegram-bot==20.3
requests~=2.30.0
yt_dlp